refactor: serve HTTP via @dcl/http-server, drop express
Replace the express + socket.io server with @dcl/http-server's Router for the
HTTP layer, keeping socket.io for WebSockets (attached to the http-server's
underlying Node http.Server via getUnderlyingServer). Removes express,
body-parser and cors.

- HTTP routes -> @dcl/http-server Router handlers (controllers/handlers/*),
  wired in service.ts (router.middleware + allowedMethods + setContext).
- Signed-fetch auth (/identities, /accounts) -> @dcl/crypto-middleware
  wellKnownComponents, preserving optional:false, the ADR-44 onError, and the
  decentraland-kernel-scene metadataValidator scene-block.
- socket.io request/recover/outcome/validation logic moved to a socket-server
  component; HTTP handlers relay to connected sockets through it.
- cors/body-size/metrics -> http-server built-ins; the manual http-metric
  recording is replaced by instrumentHttpServerWithPromClientRegistry.
- Client IP: main() stamps the Node socket's remote address onto a header
  (always overwritten so it cannot be spoofed) because the native Request does
  not expose the socket; getClientIp consults it as the lowest-priority
  fallback, preserving the prior behavior.
