+++ package.json
@@ -22,8 +22,6 @@
   },
   "devDependencies": {
     "@dcl/eslint-config": "^1.1.6",
-    "@types/body-parser": "^1.19.6",
-    "@types/express": "^5.0.6",
     "@types/jest": "^29.5.4",
     "@types/node": "^22.13.4",
     "@types/uuid": "^9.0.7",
@@ -61,11 +59,8 @@
     "@well-known-components/logger": "^3.1.3",
     "ajv": "^8.12.0",
     "ajv-formats": "^2.1.1",
-    "body-parser": "^1.20.2",
-    "cors": "^2.8.5",
     "decentraland-crypto-fetch": "^2.0.1",
     "ethers": "^6.15.0",
-    "express": "^4.18.2",
     "socket.io": "^4.7.2",
     "sql-template-strings": "^2.2.2",
     "uuid": "^9.0.1"
