+++ src/logic/socket-server/index.ts
@@ -1,0 +1,2 @@
+export { createSocketServerComponent } from './component'
+export * from './types'
