--- src/ports/server/component.ts
@@ -1,1449 +1,0 @@
-import { createServer } from 'http'
-import * as Sentry from '@sentry/node'
-import { IBaseComponent } from '@well-known-components/interfaces'
-import bodyParser from 'body-parser'
-import cors from 'cors'
-import { ethers } from 'ethers'
-import express, { Request, Response } from 'express'
-import { Server, Socket } from 'socket.io'
-import { v4 as uuid } from 'uuid'
-import { Authenticator, parseEmphemeralPayload } from '@dcl/crypto'
-import { express as authMiddleware, DecentralandSignatureData } from '@dcl/crypto-middleware'
-import { AuthChain } from '@dcl/schemas'
-import { MagicRateLimitError, MagicTokenExpiredError, MagicTokenInvalidError } from '../../adapters/magic'
-import { AddressMismatchError, DidTokenReusedError, DidTokenStaleError } from '../../logic/account-deletion'
-import { isErrorWithMessage } from '../../logic/error-handling'
-import { AppComponents } from '../../types'
-import { METHOD_DCL_PERSONAL_SIGN, ONE_HOUR_IN_MILLISECONDS, MAX_BODY_SIZE } from './constants'
-import {
-  HttpOutcomeMessage,
-  IServerComponent,
-  IdentityResponse,
-  IdentityIdValidationResponse,
-  InvalidResponseMessage,
-  LiveResponseMessage,
-  MessageType,
-  OutcomeMessage,
-  OutcomeResponseMessage,
-  RecoverMessage,
-  RecoverResponseMessage,
-  RequestMessage,
-  RequestResponseMessage,
-  RequestValidationMessage,
-  RequestValidationStatusMessage,
-  AccountDeletionMetadata,
-  AccountDeletionResponse
-} from './types'
-import {
-  validateHttpOutcomeMessage,
-  validateOutcomeMessage,
-  validateRecoverMessage,
-  validateRequestMessage,
-  validateRequestValidationMessage,
-  validateIdentityId,
-  validateIdentityRequest,
-  validateCheckpointRequest,
-  validateAccountDeletionMetadata
-} from './validations'
-
-export async function createServerComponent({
-  config,
-  logs,
-  metrics,
-  onboarding,
-  accountDeletion,
-  email,
-  nudgeJob,
-  storage,
-  tracer,
-  requestExpirationInSeconds,
-  dclPersonalSignExpirationInSeconds
-}: Pick<AppComponents, 'config' | 'logs' | 'metrics' | 'onboarding' | 'accountDeletion' | 'email' | 'nudgeJob' | 'storage' | 'tracer'> & {
-  requestExpirationInSeconds: number
-  dclPersonalSignExpirationInSeconds: number
-}): Promise<IServerComponent> {
-  // Wraps the callback function on messages to type the message that is being sent
-  const sendResponse = <T>(res: Response, statusCode: number, msg: T) => {
-    res.status(statusCode).json(msg)
-  }
-
-  const port = await config.requireNumber('HTTP_SERVER_PORT')
-  const logger = logs.getLogger('websocket-server')
-
-  const corsOptions = {
-    origin: (await config.requireString('CORS_ORIGIN')).split(';').map(origin => new RegExp(origin)),
-    methods: await config.requireString('CORS_METHODS')
-  }
-
-  // Exact-match allowlist of browser Origins permitted to call the account
-  // deletion endpoint (defense-in-depth on top of CORS). Empty disables the check.
-  const accountDeletionAllowedOrigins = new Set(
-    ((await config.getString('ACCOUNT_DELETION_ALLOWED_ORIGINS')) || '')
-      .split(';')
-      .map(origin => origin.trim().toLowerCase())
-      .filter(origin => origin.length > 0)
-  )
-
-  // Metrics configuration
-  const metricsPath = (await config.getString('WKC_METRICS_PUBLIC_PATH')) || '/metrics'
-  const metricsBearerToken = await config.getString('WKC_METRICS_BEARER_TOKEN')
-  const onboardingApiKey = await config.requireString('ONBOARDING_API_KEY')
-
-  const sockets: Record<string, Socket> = {}
-
-  let server: Server | null = null
-
-  const onConnection = (socket: Socket) =>
-    tracer.span('websocket-connection', () => {
-      // Do some work here
-      logger.log('Connected')
-      sockets[socket.id] = socket
-
-      const parentTracingContext = tracer.getTrace()
-
-      socket.on('disconnect', () =>
-        tracer.span(
-          'websocket-disconnect',
-          () => {
-            logger.log('Disconnected')
-            // Don't delete requests on disconnect — let them expire naturally via TTL.
-            // This prevents the race condition where the user is still completing
-            // the auth flow when the game client's socket temporarily disconnects.
-            delete sockets[socket.id]
-          },
-          parentTracingContext
-        )
-      )
-
-      // Wraps the callback function on messages to type the message that is being sent.
-      // On the client, the response will be received using socket.emitWithAck().
-      const ack = <T>(cb: (...args: unknown[]) => void, msg: T) =>
-        tracer.span(
-          'websocket-ack',
-          () => {
-            try {
-              cb(msg)
-            } catch (e) {
-              // This might happen if the request was done with socket.emit instead of socket.emitWithAck.
-              logger.error(`There was an error sending the response message: ${isErrorWithMessage(e) ? e.message : 'Unknown error'}`)
-            }
-          },
-          parentTracingContext
-        )
-
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      socket.on(MessageType.REQUEST, async (data: any, cb) =>
-        tracer
-          .span(
-            'websocket-request',
-            async () => {
-              let msg: RequestMessage
-              logger.log('Received a request')
-
-              try {
-                msg = validateRequestMessage(data)
-              } catch (e) {
-                ack<InvalidResponseMessage>(cb, {
-                  error: isErrorWithMessage(e) ? e.message : 'Unknown error'
-                })
-                logger.log('Received a request with invalid message')
-
-                return
-              }
-
-              let sender: string | undefined
-
-              if (msg.method !== METHOD_DCL_PERSONAL_SIGN) {
-                const authChain = msg.authChain
-
-                if (!authChain) {
-                  ack<InvalidResponseMessage>(cb, {
-                    error: 'Auth chain is required'
-                  })
-                  logger.log('Received a request without an auth chain')
-                  return
-                }
-
-                sender = Authenticator.ownerAddress(authChain)
-
-                let finalAuthority: string
-
-                try {
-                  finalAuthority = parseEmphemeralPayload(authChain[authChain.length - 1].payload).ephemeralAddress
-                } catch (e) {
-                  ack<InvalidResponseMessage>(cb, {
-                    error: 'Could not get final authority from auth chain'
-                  })
-                  logger.log('Received a request with invalid auth chain')
-                  return
-                }
-
-                const validationResult = await Authenticator.validateSignature(finalAuthority, authChain, null)
-
-                if (!validationResult.ok) {
-                  ack<InvalidResponseMessage>(cb, {
-                    error: validationResult.message ?? 'Signature validation failed'
-                  })
-
-                  logger.log('Received a request with invalid signature')
-                  return
-                }
-              }
-
-              const requestId = uuid()
-              const expiration = new Date(
-                Date.now() +
-                  (msg.method !== METHOD_DCL_PERSONAL_SIGN ? requestExpirationInSeconds : dclPersonalSignExpirationInSeconds) * 1000
-              )
-              const code = Math.floor(Math.random() * 100)
-
-              await storage.setRequest(requestId, {
-                requestId: requestId,
-                socketId: socket.id,
-                requiresValidation: false,
-                expiration,
-                code,
-                method: msg.method,
-                params: msg.params,
-                sender: sender?.toLowerCase()
-              })
-
-              ack<RequestResponseMessage>(cb, {
-                requestId,
-                expiration,
-                code
-              })
-
-              logger.log(`[METHOD:${msg.method}][RID:${requestId}][EXP:${expiration.getTime()}] Successfully registered request response`)
-            },
-            parentTracingContext
-          )
-          .catch(e => {
-            Sentry.captureException(e)
-            logger.error(`Unexpected error in ${MessageType.REQUEST} handler: ${isErrorWithMessage(e) ? e.message : 'Unknown error'}`)
-            ack<InvalidResponseMessage>(cb, { error: 'Internal server error' })
-          })
-      )
-
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      socket.on(MessageType.RECOVER, (data: any, cb) =>
-        tracer
-          .span(
-            'websocket-recover',
-            async () => {
-              let msg: RecoverMessage
-              logger.log('Received a recover request')
-
-              try {
-                msg = validateRecoverMessage(data)
-              } catch (e) {
-                ack<InvalidResponseMessage>(cb, {
-                  error: isErrorWithMessage(e) ? e.message : 'Unknown error'
-                })
-
-                logger.log('Received a recover request with invalid message')
-                return
-              }
-
-              const request = await storage.getRequest(msg.requestId)
-
-              if (!request) {
-                ack<InvalidResponseMessage>(cb, {
-                  error: `Request with id "${msg.requestId}" not found`
-                })
-
-                logger.log(`[RID:${msg.requestId}] Received a recover request for a non-existent request`)
-
-                return
-              }
-
-              if (request.fulfilled) {
-                ack<InvalidResponseMessage>(cb, {
-                  error: `Request with id "${msg.requestId}" has already been fulfilled`
-                })
-
-                logger.log(`[RID:${msg.requestId}] Received a recover request for an already fulfilled request`)
-
-                return
-              }
-
-              if (request.expiration < new Date()) {
-                await storage.setRequest(msg.requestId, null)
-
-                ack<InvalidResponseMessage>(cb, {
-                  error: `Request with id "${msg.requestId}" has expired`
-                })
-
-                logger.log(`[RID:${msg.requestId}] Received a recover request for an expired request`)
-
-                return
-              }
-
-              ack<RecoverResponseMessage>(cb, {
-                expiration: request.expiration,
-                code: request.code,
-                method: request.method,
-                params: request.params,
-                sender: request.sender
-              })
-
-              logger.log(
-                `[METHOD:${request.method}][RID:${request.requestId}][EXP:${request.expiration.getTime()}] Successfully recovered request`
-              )
-            },
-            parentTracingContext
-          )
-          .catch(e => {
-            Sentry.captureException(e)
-            logger.error(`Unexpected error in ${MessageType.RECOVER} handler: ${isErrorWithMessage(e) ? e.message : 'Unknown error'}`)
-            ack<InvalidResponseMessage>(cb, { error: 'Internal server error' })
-          })
-      )
-
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      socket.on(MessageType.OUTCOME, (data: any, cb) =>
-        tracer
-          .span(
-            'websocket-outcome',
-            async () => {
-              let msg: OutcomeMessage
-
-              try {
-                msg = validateOutcomeMessage(data)
-              } catch (e) {
-                ack<InvalidResponseMessage>(cb, {
-                  error: isErrorWithMessage(e) ? e.message : 'Unknown error'
-                })
-
-                logger.log('Received an outcome message with invalid message')
-
-                return
-              }
-
-              const request = await storage.getRequest(msg.requestId)
-
-              if (!request) {
-                ack<InvalidResponseMessage>(cb, {
-                  error: `Request with id "${msg.requestId}" not found`
-                })
-
-                logger.log(`[RID:${msg.requestId}] Received an outcome message for a non-existent request`)
-
-                return
-              }
-
-              if (request.fulfilled) {
-                ack<InvalidResponseMessage>(cb, {
-                  error: `Request with id "${msg.requestId}" has already been fulfilled`
-                })
-
-                logger.log(`[RID:${msg.requestId}] Received an outcome message for an already fulfilled request`)
-
-                return
-              }
-
-              if (request.response) {
-                ack<InvalidResponseMessage>(cb, {
-                  error: `Request with id "${msg.requestId}" already has a response`
-                })
-
-                logger.log(`[RID:${msg.requestId}] Received an outcome message for a request that already has a response`)
-
-                return
-              }
-
-              if (request.expiration < new Date()) {
-                await storage.setRequest(msg.requestId, null)
-
-                ack<InvalidResponseMessage>(cb, {
-                  error: `Request with id "${msg.requestId}" has expired`
-                })
-
-                logger.log(`[RID:${msg.requestId}] Received an outcome message for an expired request`)
-
-                return
-              }
-
-              const outcomeMessage: OutcomeResponseMessage = msg
-
-              // If it has a socketId and the socket is still connected, send via socket.
-              // Otherwise, store the response for polling via GET /requests/:requestId.
-              if (request.socketId && sockets[request.socketId]) {
-                const storedSocket = sockets[request.socketId]
-
-                // Mark as fulfilled instead of deleting — allows frontend to distinguish "consumed" from "never existed"
-                await storage.setRequest(msg.requestId, {
-                  requestId: msg.requestId,
-                  socketId: request.socketId,
-                  fulfilled: true,
-                  expiration: request.expiration,
-                  code: 0,
-                  method: '',
-                  params: [],
-                  requiresValidation: false
-                })
-
-                storedSocket.emit(MessageType.OUTCOME, outcomeMessage)
-                logger.log(
-                  `[METHOD:${request.method}][RID:${
-                    request.requestId
-                  }][EXP:${request.expiration.getTime()}] Successfully sent outcome message to the client`
-                )
-              } else {
-                // Socket gone or HTTP-created request — persist response for polling
-                await storage.setRequest(msg.requestId, {
-                  ...request,
-                  response: outcomeMessage
-                })
-                logger.log(
-                  `[METHOD:${request.method}][RID:${
-                    request.requestId
-                  }][EXP:${request.expiration.getTime()}] Stored outcome for polling (socket unavailable)`
-                )
-              }
-
-              ack<object>(cb, {})
-            },
-            parentTracingContext
-          )
-          .catch(e => {
-            Sentry.captureException(e)
-            logger.error(`Unexpected error in ${MessageType.OUTCOME} handler: ${isErrorWithMessage(e) ? e.message : 'Unknown error'}`)
-            ack<InvalidResponseMessage>(cb, { error: 'Internal server error' })
-          })
-      )
-
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      socket.on(MessageType.REQUEST_VALIDATION_STATUS, (data: any, cb) => {
-        tracer
-          .span(
-            'websocket-request-validation',
-            async () => {
-              let msg: RequestValidationMessage
-
-              try {
-                msg = validateRequestValidationMessage(data)
-              } catch (e) {
-                logger.log('Received an outcome message with invalid message')
-                return ack<InvalidResponseMessage>(cb, {
-                  error: isErrorWithMessage(e) ? e.message : 'Unknown error'
-                })
-              }
-
-              const request = await storage.getRequest(msg.requestId)
-
-              if (!request) {
-                logger.log(`[RID:${msg.requestId}] Tried to communicate that the request must be validated but it doesn't exist`)
-                return ack<InvalidResponseMessage>(cb, {
-                  error: `Request with id "${msg.requestId}" not found`
-                })
-              }
-
-              if (request.fulfilled) {
-                logger.log(
-                  `[RID:${msg.requestId}] Tried to communicate that the request must be validated but it has already been fulfilled`
-                )
-                return ack<InvalidResponseMessage>(cb, {
-                  error: `Request with id "${msg.requestId}" has already been fulfilled`
-                })
-              }
-
-              if (request.expiration < new Date()) {
-                await storage.setRequest(msg.requestId, null)
-
-                logger.log(`[RID:${msg.requestId}] Tried to communicate that the request must be validated but it has expired`)
-                return ack<InvalidResponseMessage>(cb, {
-                  error: `Request with id "${msg.requestId}" has expired`
-                })
-              }
-
-              if (request.socketId && sockets[request.socketId] && !request.requiresValidation) {
-                const storedSocket = sockets[request.socketId]
-
-                // Relay the request validation to the client
-                storedSocket.emit(MessageType.REQUEST_VALIDATION_STATUS, { requestId: msg.requestId, code: request.code })
-              }
-
-              request.requiresValidation = true
-
-              ack<object>(cb, {})
-            },
-            parentTracingContext
-          )
-          .catch(e => {
-            Sentry.captureException(e)
-            logger.error(
-              `Unexpected error in ${MessageType.REQUEST_VALIDATION_STATUS} handler: ${isErrorWithMessage(e) ? e.message : 'Unknown error'}`
-            )
-            ack<InvalidResponseMessage>(cb, { error: 'Internal server error' })
-          })
-      })
-    })
-
-  const start: IBaseComponent['start'] = async () => {
-    if (server) {
-      return
-    }
-    const logger = logs.getLogger('websocket-server')
-    const identityLogger = logs.getLogger('identity-endpoints')
-
-    logger.log('Starting socket server...')
-
-    const app = express()
-    const httpServer = createServer(app)
-
-    // Middleware to parse JSON in the request body
-    app.use(bodyParser.json({ limit: MAX_BODY_SIZE }))
-    app.use(cors(corsOptions))
-
-    // Request tracking middleware
-    app.use((req, res, next) => {
-      if (req.path === metricsPath) {
-        return next()
-      }
-
-      const labels = {
-        method: req.method,
-        handler: '',
-        code: 200
-      }
-
-      const startTimerResult = metrics.startTimer('http_request_duration_seconds', labels)
-      // eslint-disable-next-line @typescript-eslint/no-empty-function
-      const end = startTimerResult?.end || (() => {})
-
-      res.on('finish', () => {
-        labels.code = res.statusCode
-        if (req.route?.path) {
-          labels.handler = req.baseUrl + req.route.path
-        }
-
-        const contentLength = parseInt(req.get('content-length') || '0', 10) || 0
-        metrics.observe('http_request_size_bytes', labels, contentLength)
-        metrics.increment('http_requests_total', labels)
-        end(labels)
-      })
-
-      next()
-    })
-
-    app.get('/health/ready', (_req, res) => {
-      res.sendStatus(200)
-    })
-
-    app.get('/health/startup', (_req, res) => {
-      res.sendStatus(200)
-    })
-
-    app.get('/health/live', (_req, res) => {
-      return sendResponse<LiveResponseMessage>(res, 200, {
-        timestamp: Date.now()
-      })
-    })
-
-    // Helper function to validate auth chain
-    const validateAuthChain = async (authChain: AuthChain): Promise<{ sender: string; finalAuthority: string }> => {
-      if (!authChain.length) {
-        throw new Error('Auth chain is required')
-      }
-
-      const sender = Authenticator.ownerAddress(authChain)
-
-      let finalAuthority: string
-
-      try {
-        const ephemeralPayload = parseEmphemeralPayload(authChain[authChain.length - 1].payload)
-
-        finalAuthority = ephemeralPayload.ephemeralAddress
-      } catch (e) {
-        if (e instanceof Error && e.message === 'Ephemeral payload has expired') {
-          throw e
-        }
-        throw new Error('Could not get final authority from auth chain')
-      }
-
-      const validationResult = await Authenticator.validateSignature(finalAuthority, authChain, null)
-
-      if (!validationResult.ok) {
-        throw new Error(validationResult.message ?? 'Signature validation failed')
-      }
-
-      return { sender, finalAuthority }
-    }
-
-    // Helper function to normalize IP address (convert IPv6-mapped IPv4 to IPv4)
-    const normalizeIp = (ip: string): string => {
-      // Convert IPv6-mapped IPv4 (::ffff:xxx.xxx.xxx.xxx) to IPv4
-      if (ip.startsWith('::ffff:')) {
-        return ip.substring(7)
-      }
-      return ip.trim()
-    }
-
-    // Helper function to get client IP address from request
-    // Prioritizes trusted headers (True-Client-IP, X-Real-IP, Cloudflare's CF-Connecting-IP) over X-Forwarded-For
-    const getClientIp = (req: Request): string => {
-      // Check True-Client-IP header (set by proxies like Cloudflare, contains the visitor's IP address)
-      const trueClientIp = req.headers['true-client-ip']
-      if (trueClientIp) {
-        const ip = Array.isArray(trueClientIp) ? trueClientIp[0] : trueClientIp
-        return normalizeIp(ip)
-      }
-
-      // Check X-Real-IP header (set by proxies when configured, more trustworthy than X-Forwarded-For)
-      const xRealIp = req.headers['x-real-ip']
-      if (xRealIp) {
-        const ip = Array.isArray(xRealIp) ? xRealIp[0] : xRealIp
-        return normalizeIp(ip)
-      }
-
-      // Check CF-Connecting-IP header first (Cloudflare's trusted header, cannot be spoofed)
-      const cfConnectingIp = req.headers['cf-connecting-ip']
-      if (cfConnectingIp) {
-        const ip = Array.isArray(cfConnectingIp) ? cfConnectingIp[0] : cfConnectingIp
-        return normalizeIp(ip)
-      }
-
-      // Check X-Forwarded-For header (can be spoofed, use with caution)
-      // Take the first IP in the chain (original client)
-      const xForwardedFor = req.headers['x-forwarded-for']
-      if (xForwardedFor) {
-        const ips = Array.isArray(xForwardedFor) ? xForwardedFor[0] : xForwardedFor.split(',')[0]
-        return normalizeIp(ips)
-      }
-
-      // Fallback to req.ip (requires express trust proxy configuration) or connection remote address
-      const fallbackIp = req.ip || req.socket.remoteAddress || 'unknown'
-      return normalizeIp(fallbackIp)
-    }
-
-    // Helper function to check if two IPs match, considering subnet/region matching
-    // For IPv4, this can match by subnet (e.g., 10.0.16.* matches 10.0.16.*)
-    const ipsMatch = (ip1: string, ip2: string): boolean => {
-      if (!ip1 || !ip2 || ip1 === 'unknown' || ip2 === 'unknown') {
-        return false
-      }
-
-      // Exact match
-      if (ip1 === ip2) {
-        return true
-      }
-
-      // Normalize both IPs
-      const normalizedIp1 = normalizeIp(ip1)
-      const normalizedIp2 = normalizeIp(ip2)
-
-      // Exact match after normalization
-      if (normalizedIp1 === normalizedIp2) {
-        return true
-      }
-
-      // IPv4 subnet matching: check if they're in the same /24 subnet (first 3 octets)
-      // This helps with VPNs that might use different edge servers but same region
-      const ipv4Regex = /^(\d{1,3})\.(\d{1,3})\.(\d{1,3})\.(\d{1,3})$/
-      const match1 = normalizedIp1.match(ipv4Regex)
-      const match2 = normalizedIp2.match(ipv4Regex)
-
-      if (match1 && match2) {
-        // Match by /24 subnet (first 3 octets)
-        if (match1[1] === match2[1] && match1[2] === match2[2] && match1[3] === match2[3]) {
-          return true
-        }
-      }
-
-      return false
-    }
-
-    // Helper function to format IP-related headers for logging
-    const formatIpHeaders = (req: Request): string => {
-      return `true-client-ip=${req.headers['true-client-ip']}, x-real-ip=${req.headers['x-real-ip']}, cf-connecting-ip=${req.headers['cf-connecting-ip']}, x-forwarded-for=${req.headers['x-forwarded-for']}`
-    }
-
-    app.post('/requests', async (req: Request, res: Response) => {
-      const data = req.body
-      let msg: RequestMessage
-
-      try {
-        msg = validateRequestMessage(data)
-      } catch (e) {
-        return sendResponse<InvalidResponseMessage>(res, 400, {
-          error: isErrorWithMessage(e) ? e.message : 'Unknown error'
-        })
-      }
-
-      let sender: string | undefined
-
-      if (msg.method !== METHOD_DCL_PERSONAL_SIGN) {
-        try {
-          const { sender: validatedSender } = await validateAuthChain(msg.authChain || [])
-          sender = validatedSender
-        } catch (e) {
-          return sendResponse<InvalidResponseMessage>(res, 400, {
-            error: isErrorWithMessage(e) ? e.message : 'Unknown error'
-          })
-        }
-      }
-
-      const requestId = uuid()
-      const expiration = new Date(
-        Date.now() + (msg.method !== METHOD_DCL_PERSONAL_SIGN ? requestExpirationInSeconds : dclPersonalSignExpirationInSeconds) * 1000
-      )
-      const code = Math.floor(Math.random() * 100)
-
-      await storage.setRequest(requestId, {
-        requestId: requestId,
-        expiration,
-        code,
-        method: msg.method,
-        params: msg.params,
-        sender: sender?.toLowerCase(),
-        requiresValidation: false
-      })
-
-      sendResponse<RequestResponseMessage>(res, 201, {
-        requestId,
-        expiration,
-        code
-      })
-    })
-
-    // Get a request by id
-    app.get('/v2/requests/:requestId', async (req: Request, res: Response) => {
-      const requestId = req.params.requestId
-      const request = await storage.getRequest(requestId)
-
-      if (!request) {
-        return sendResponse<InvalidResponseMessage>(res, 404, {
-          error: `Request with id "${requestId}" not found`
-        })
-      }
-
-      if (request.fulfilled) {
-        return sendResponse<InvalidResponseMessage>(res, 410, {
-          error: `Request with id "${requestId}" has already been fulfilled`
-        })
-      }
-
-      if (request.expiration < new Date()) {
-        await storage.setRequest(requestId, null)
-
-        return sendResponse<InvalidResponseMessage>(res, 410, {
-          error: `Request with id "${requestId}" has expired`
-        })
-      }
-
-      sendResponse<RecoverResponseMessage>(res, 200, {
-        expiration: request.expiration,
-        code: request.code,
-        method: request.method,
-        params: request.params,
-        sender: request.sender
-      })
-    })
-
-    // Communicate that the request must be validated
-    app.post('/v2/requests/:requestId/validation', async (req: Request, res: Response) => {
-      const requestId = req.params.requestId
-
-      const request = await storage.getRequest(requestId)
-
-      if (!request) {
-        logger.log(`[RID:${requestId}] Received a validation request message for a non-existent request`)
-        return sendResponse<InvalidResponseMessage>(res, 404, {
-          error: `Request with id "${requestId}" not found`
-        })
-      }
-
-      if (request.fulfilled) {
-        logger.log(`[RID:${requestId}] Received a validation request message for an already fulfilled request`)
-        return sendResponse<InvalidResponseMessage>(res, 410, {
-          error: `Request with id "${requestId}" has already been fulfilled`
-        })
-      }
-
-      if (request.expiration < new Date()) {
-        logger.log(`[RID:${requestId}] Received a validation request message for an expired request`)
-
-        // Remove the request from the storage as it is expired
-        await storage.setRequest(requestId, null)
-
-        return sendResponse<InvalidResponseMessage>(res, 410, {
-          error: `Request with id "${requestId}" has expired`
-        })
-      }
-
-      if (request.socketId && sockets[request.socketId] && !request.requiresValidation) {
-        const storedSocket = sockets[request.socketId]
-
-        logger.log(`[RID:${requestId}] Successfully sent request validation to the client via socket`)
-        // Send the request validation to the client
-        storedSocket.emit(MessageType.REQUEST_VALIDATION_STATUS, { requestId })
-      }
-
-      request.requiresValidation = true
-
-      res.sendStatus(204)
-    })
-
-    // Get the request validation status
-    app.get('/v2/requests/:requestId/validation', async (req: Request, res: Response) => {
-      const requestId = req.params.requestId
-      const request = await storage.getRequest(requestId)
-
-      if (!request) {
-        return sendResponse<InvalidResponseMessage>(res, 404, {
-          error: `Request with id "${requestId}" not found`
-        })
-      }
-
-      if (request.fulfilled) {
-        return sendResponse<InvalidResponseMessage>(res, 410, {
-          error: `Request with id "${requestId}" has already been fulfilled`
-        })
-      }
-
-      if (request.expiration < new Date()) {
-        await storage.setRequest(requestId, null)
-
-        return sendResponse<InvalidResponseMessage>(res, 410, {
-          error: `Request with id "${requestId}" has expired`
-        })
-      }
-
-      sendResponse<RequestValidationStatusMessage>(res, 200, { requiresValidation: request.requiresValidation })
-    })
-
-    // Get the outcome of a request
-    app.get('/requests/:requestId', async (req: Request, res: Response) => {
-      const requestId = req.params.requestId
-      const request = await storage.getRequest(requestId)
-
-      if (!request) {
-        return sendResponse<InvalidResponseMessage>(res, 404, {
-          error: `Request with id "${requestId}" not found`
-        })
-      }
-
-      if (request.fulfilled) {
-        return sendResponse<InvalidResponseMessage>(res, 410, {
-          error: `Request with id "${requestId}" has already been fulfilled`
-        })
-      }
-
-      if (request.expiration < new Date()) {
-        await storage.setRequest(requestId, null)
-        return sendResponse<InvalidResponseMessage>(res, 410, {
-          error: `Request with id "${requestId}" has expired`
-        })
-      }
-
-      if (!request.response) {
-        return sendResponse<InvalidResponseMessage>(res, 204, {
-          error: `Request with id "${requestId}" has not been completed`
-        })
-      }
-
-      logger.log(`[RID:${requestId}] Successfully sent outcome message to the client via HTTP`)
-
-      // Mark as fulfilled instead of deleting — allows frontend to distinguish "consumed" from "never existed"
-      await storage.setRequest(requestId, {
-        requestId,
-        fulfilled: true,
-        expiration: request.expiration,
-        code: 0,
-        method: '',
-        params: [],
-        requiresValidation: false
-      })
-      sendResponse<OutcomeResponseMessage>(res, 200, request.response)
-    })
-
-    // Record the outcome of a request
-    app.post('/v2/requests/:requestId/outcome', async (req: Request, res: Response) => {
-      const requestId = req.params.requestId
-
-      const data = req.body
-      let msg: HttpOutcomeMessage
-
-      try {
-        msg = validateHttpOutcomeMessage(data)
-      } catch (e) {
-        return sendResponse<InvalidResponseMessage>(res, 400, {
-          error: isErrorWithMessage(e) ? e.message : 'Unknown error'
-        })
-      }
-
-      const request = await storage.getRequest(requestId)
-
-      if (!request) {
-        logger.log(`[RID:${requestId}] Received an outcome message for a non-existent request`)
-        return sendResponse<InvalidResponseMessage>(res, 404, {
-          error: `Request with id "${requestId}" not found`
-        })
-      }
-
-      if (request.fulfilled) {
-        logger.log(`[RID:${requestId}] Received an outcome message for an already fulfilled request`)
-        return sendResponse<InvalidResponseMessage>(res, 410, {
-          error: `Request with id "${requestId}" has already been fulfilled`
-        })
-      }
-
-      if (request.response) {
-        logger.log(`[RID:${requestId}] Received an outcome message for a request that already has a response`)
-
-        return sendResponse<InvalidResponseMessage>(res, 400, {
-          error: `Request with id "${requestId}" already has a response`
-        })
-      }
-
-      if (request.expiration < new Date()) {
-        await storage.setRequest(requestId, null)
-
-        logger.log(`[RID:${requestId}] Received an outcome message for an expired request`)
-        return sendResponse<InvalidResponseMessage>(res, 410, {
-          error: `Request with id "${requestId}" has expired`
-        })
-      }
-
-      const outcomeMessage: OutcomeResponseMessage = {
-        ...msg,
-        requestId
-      }
-
-      if (request.socketId && sockets[request.socketId]) {
-        const storedSocket = sockets[request.socketId]
-        storedSocket.emit(MessageType.OUTCOME, outcomeMessage)
-        logger.log(
-          `[METHOD:${request.method}][RID:${
-            request.requestId
-          }][EXP:${request.expiration.getTime()}] Successfully sent outcome message to the client via socket`
-        )
-        // Mark as fulfilled instead of deleting — allows frontend to distinguish "consumed" from "never existed"
-        await storage.setRequest(requestId, {
-          requestId,
-          socketId: request.socketId,
-          fulfilled: true,
-          expiration: request.expiration,
-          code: 0,
-          method: '',
-          params: [],
-          requiresValidation: false
-        })
-      } else {
-        // Socket gone or HTTP-created request — persist response for polling via GET /requests/:requestId
-        await storage.setRequest(requestId, {
-          ...request,
-          response: outcomeMessage
-        })
-        logger.log(
-          `[METHOD:${request.method}][RID:${
-            request.requestId
-          }][EXP:${request.expiration.getTime()}] Stored outcome for polling (socket unavailable)`
-        )
-      }
-
-      return res.sendStatus(200)
-    })
-
-    // Store identity endpoint with signed fetch middleware validation
-    app.post(
-      '/identities',
-      authMiddleware({
-        optional: false,
-        onError: err => ({
-          error: err.message,
-          message: 'This endpoint requires a signed fetch request. See ADR-44.'
-        }),
-        metadataValidator: metadata => metadata?.signer !== 'decentraland-kernel-scene' // prevent requests from scenes
-      }),
-      async (req: Request & DecentralandSignatureData) => {
-        const res = req.res as Response
-        identityLogger.log('Received a request to create identity')
-        try {
-          const { identity, isMobile } = validateIdentityRequest(req.body)
-
-          if (!identity) {
-            identityLogger.log('Received a request to create identity without AuthIdentity in body')
-            return sendResponse<InvalidResponseMessage>(res, 400, {
-              error: 'AuthIdentity is required in request body'
-            })
-          }
-
-          // Validate auth chain using the same logic as /requests endpoint
-          let identitySender: string
-          try {
-            const { sender, finalAuthority } = await validateAuthChain(identity.authChain)
-            identitySender = sender
-            // Verify that the ephemeral wallet address matches the finalAuthority from auth chain
-            if (identity.ephemeralIdentity.address.toLowerCase() !== finalAuthority.toLowerCase()) {
-              identityLogger.log(`Ephemeral wallet address does not match auth chain final authority for sender: ${identitySender}`)
-              return sendResponse<InvalidResponseMessage>(res, 403, {
-                error: 'Ephemeral wallet address does not match auth chain final authority'
-              })
-            }
-
-            // Verify that the user making the request is the same as the one who signed the identity
-            const requestSender = req.auth
-            if (!requestSender || requestSender.toLowerCase() !== identitySender.toLowerCase()) {
-              identityLogger.log(`Request sender (${requestSender}) does not match identity owner (${identitySender})`)
-              return sendResponse<InvalidResponseMessage>(res, 403, {
-                error: 'Request sender does not match identity owner'
-              })
-            }
-
-            const wallet = new ethers.Wallet(identity.ephemeralIdentity.privateKey)
-
-            if (wallet.address.toLowerCase() !== identity.ephemeralIdentity.address.toLowerCase()) {
-              identityLogger.log(`Ephemeral private key does not match the provided address for sender: ${identitySender}`)
-              return sendResponse<InvalidResponseMessage>(res, 403, {
-                error: 'Ephemeral private key does not match the provided address'
-              })
-            }
-          } catch (e) {
-            const errorMessage = isErrorWithMessage(e) ? e.message : 'Unknown error'
-            identityLogger.log(`Received a request to create identity with invalid auth chain: ${errorMessage}`)
-            return sendResponse<InvalidResponseMessage>(res, 400, {
-              error: errorMessage
-            })
-          }
-
-          const identityId = uuid()
-          // Always use 15 minutes expiration for storage (controls when identity is removed from storage)
-          const storageExpiration = new Date(Date.now() + ONE_HOUR_IN_MILLISECONDS)
-          const clientIp = getClientIp(req)
-
-          await storage.setIdentity(identityId, {
-            identityId,
-            identity,
-            expiration: storageExpiration,
-            createdAt: new Date(),
-            ipAddress: clientIp,
-            isMobile: isMobile === true
-          })
-
-          await storage.setIdentityStatus(identityId, {
-            expiration: storageExpiration,
-            createdAt: new Date(),
-            consumed: false,
-            signer: identitySender
-          })
-
-          identityLogger.log(
-            `[IID:${identityId}][SIGNER:${identitySender}][EXP:${storageExpiration.getTime()}][Mobile:${
-              isMobile === true
-            }] Successfully created identity from IP: ${clientIp}. Headers: ${formatIpHeaders(req)}`
-          )
-
-          sendResponse<IdentityResponse>(res, 201, {
-            identityId,
-            expiration: storageExpiration
-          })
-        } catch (e) {
-          const errorMessage = isErrorWithMessage(e) ? e.message : 'Unknown error'
-          identityLogger.log(`Received a request to create identity with invalid message: ${errorMessage}`)
-          return sendResponse<InvalidResponseMessage>(res, 400, {
-            error: errorMessage
-          })
-        }
-      }
-    )
-
-    // Account deletion endpoint — permanently deletes the user's Magic account
-    // (irreversible) and purges local PII. Layered auth: DCL signed-fetch
-    // (req.auth) + a fresh Magic DID token, cross-checked to the same address.
-    app.delete(
-      '/accounts',
-      authMiddleware({
-        optional: false,
-        onError: err => ({
-          error: err.message,
-          message: 'This endpoint requires a signed fetch request. See ADR-44.'
-        }),
-        metadataValidator: metadata => metadata?.signer !== 'decentraland-kernel-scene' // prevent requests from scenes
-      }),
-      async (req: Request & DecentralandSignatureData) => {
-        const res = req.res as Response
-        const deletionLogger = logs.getLogger('account-deletion-endpoint')
-
-        // Defense-in-depth: restrict to official Decentraland browser origins.
-        const origin = (req.headers.origin || '').toLowerCase()
-        if (accountDeletionAllowedOrigins.size > 0 && (!origin || !accountDeletionAllowedOrigins.has(origin))) {
-          deletionLogger.log(`Rejected account deletion from disallowed origin: ${req.headers.origin ?? 'none'}`)
-          return sendResponse<InvalidResponseMessage>(res, 403, { error: 'Origin not allowed' })
-        }
-
-        // The DID token travels in the signed-fetch metadata, so it is covered by
-        // the request signature (method:path:timestamp:metadata) — no request body.
-        let metadata: AccountDeletionMetadata
-        try {
-          metadata = validateAccountDeletionMetadata(req.authMetadata)
-        } catch (e) {
-          return sendResponse<InvalidResponseMessage>(res, 400, {
-            error: isErrorWithMessage(e) ? e.message : 'Unknown error'
-          })
-        }
-
-        const requestSender = req.auth
-        if (!requestSender) {
-          return sendResponse<InvalidResponseMessage>(res, 401, { error: 'Request signer is required' })
-        }
-
-        try {
-          const result = await accountDeletion.deleteAccount({
-            signedFetchAddress: requestSender,
-            didToken: metadata.didToken,
-            ip: getClientIp(req)
-          })
-
-          return sendResponse<AccountDeletionResponse>(res, 200, {
-            deleted: true,
-            address: result.address,
-            magic: result.magic
-          })
-        } catch (e) {
-          const message = isErrorWithMessage(e) ? e.message : 'Unknown error'
-
-          if (e instanceof MagicTokenInvalidError || e instanceof MagicTokenExpiredError) {
-            deletionLogger.log(`Rejected account deletion: invalid DID token: ${message}`)
-            return sendResponse<InvalidResponseMessage>(res, 401, { error: message })
-          }
-
-          if (e instanceof AddressMismatchError || e instanceof DidTokenStaleError || e instanceof DidTokenReusedError) {
-            deletionLogger.log(`Rejected account deletion: ${message}`)
-            return sendResponse<InvalidResponseMessage>(res, 403, { error: message })
-          }
-
-          if (e instanceof MagicRateLimitError) {
-            deletionLogger.log(`Account deletion rate limited by Magic: ${message}`)
-            return sendResponse<InvalidResponseMessage>(res, 429, { error: message })
-          }
-
-          // MagicAuthError (our misconfig), MagicDeletionError, and anything else.
-          deletionLogger.error(`Account deletion failed: ${message}`)
-          return sendResponse<InvalidResponseMessage>(res, 500, { error: 'Internal server error' })
-        }
-      }
-    )
-
-    // identities validation endpoint - returns identity for auto-login
-    app.get('/identities/:id', async (req: Request, res: Response) => {
-      const identityId = req.params.id
-      identityLogger.log(`Received a request to retrieve identity: ${identityId}`)
-
-      if (!validateIdentityId(identityId)) {
-        identityLogger.log(`[IID:${identityId}] Received a request to retrieve identity with invalid format`)
-        return sendResponse<InvalidResponseMessage>(res, 400, {
-          error: 'Invalid identity format'
-        })
-      }
-
-      const identity = await storage.getIdentity(identityId)
-
-      if (!identity) {
-        const status = await storage.getIdentityStatus(identityId)
-        if (status) {
-          if (status.deletionReason === 'consumed') {
-            identityLogger.log(`[IID:${identityId}][SIGNER:${status.signer}] Received a request to retrieve an already consumed identity`)
-            return sendResponse<InvalidResponseMessage>(res, 404, {
-              error: 'Identity was already consumed'
-            })
-          }
-          if (status.deletionReason === 'expired') {
-            identityLogger.log(`[IID:${identityId}][SIGNER:${status.signer}] Received a request to retrieve an expired identity`)
-            return sendResponse<InvalidResponseMessage>(res, 404, {
-              error: 'Identity has expired'
-            })
-          }
-          if (status.deletionReason === 'ip_mismatch') {
-            identityLogger.log(
-              `[IID:${identityId}][SIGNER:${status.signer}] Received a request to retrieve an identity deleted due to IP mismatch`
-            )
-            return sendResponse<InvalidResponseMessage>(res, 404, {
-              error: 'Identity was deleted due to IP mismatch'
-            })
-          }
-          // Tombstone exists but no deletion reason → identity was evicted from cache by Redis TTL
-          identityLogger.log(
-            `[IID:${identityId}][SIGNER:${
-              status.signer
-            }][CREATED:${status.createdAt.toISOString()}] Received a request to retrieve an identity evicted by TTL`
-          )
-          return sendResponse(res, 404, {
-            error: 'Identity was evicted',
-            createdAt: status.createdAt.toISOString()
-          })
-        }
-        identityLogger.log(`[IID:${identityId}] Received a request to retrieve a non-existent identity`)
-        return sendResponse<InvalidResponseMessage>(res, 404, {
-          error: 'Identity not found'
-        })
-      }
-
-      const signer = Authenticator.ownerAddress(identity.identity.authChain)
-
-      if (identity.expiration < new Date()) {
-        await storage.deleteIdentity(identityId)
-        await storage.updateIdentityStatus(identityId, { consumed: false, deletionReason: 'expired' })
-        identityLogger.log(`[IID:${identityId}][SIGNER:${signer}] Received a request to retrieve an expired identity`)
-        return sendResponse<InvalidResponseMessage>(res, 410, {
-          error: 'Identity has expired'
-        })
-      }
-
-      // Validate that the IP address matches the one used when creating the identity
-      // Uses flexible matching to handle Cloudflare and VPN edge server differences
-      const clientIp = getClientIp(req)
-
-      if (identity.isMobile) {
-        // Log header details if IPs differ (for debugging), but allow the request
-        if (!ipsMatch(identity.ipAddress, clientIp)) {
-          identityLogger.log(
-            `[IID:${identityId}][SIGNER:${signer}] Mobile IP mismatch (allowed). Stored: ${
-              identity.ipAddress
-            }, Request: ${clientIp}. Headers: ${formatIpHeaders(req)}`
-          )
-        }
-        // Continue without blocking for mobile
-      } else if (!ipsMatch(identity.ipAddress, clientIp)) {
-        // Non-mobile: delete identity and return 403
-        await storage.deleteIdentity(identityId)
-        await storage.updateIdentityStatus(identityId, { consumed: false, deletionReason: 'ip_mismatch' })
-        identityLogger.log(
-          `[IID:${identityId}][SIGNER:${signer}] Received a request to retrieve identity from different IP. Stored: ${identity.ipAddress}, Request: ${clientIp}. Identity deleted.`
-        )
-        return sendResponse<InvalidResponseMessage>(res, 403, {
-          error: 'IP address mismatch'
-        })
-      }
-
-      try {
-        // Delete the identity from the storage
-        await storage.deleteIdentity(identityId)
-        await storage.updateIdentityStatus(identityId, { consumed: true, deletionReason: 'consumed' })
-
-        identityLogger.log(
-          `[IID:${identityId}][SIGNER:${signer}][EXP:${identity.expiration.getTime()}] Successfully served identity to IP: ${clientIp}`
-        )
-
-        // Return the identity for auto-login
-        sendResponse<IdentityIdValidationResponse>(res, 200, {
-          identity: identity.identity
-        })
-      } catch (error) {
-        const errorMessage = isErrorWithMessage(error) ? error.message : 'Unknown error'
-        identityLogger.error(`[IID:${identityId}][SIGNER:${signer}] Error serving identity: ${errorMessage}`)
-        return sendResponse<InvalidResponseMessage>(res, 500, {
-          error: 'Internal server error'
-        })
-      }
-    })
-
-    // Record onboarding checkpoint event
-    app.post('/onboarding/checkpoint', async (req: Request, res: Response) => {
-      const onboardingLogger = logs.getLogger('onboarding-endpoint')
-
-      const authHeader = req.headers['authorization'] ?? ''
-      const provided = authHeader.startsWith('Bearer ') ? authHeader.slice(7) : ''
-      if (provided !== onboardingApiKey) {
-        sendResponse(res, 401, { error: 'Unauthorized' })
-        return
-      }
-
-      let checkpointReq
-      try {
-        checkpointReq = validateCheckpointRequest(req.body)
-      } catch (e) {
-        const validationError = isErrorWithMessage(e) ? e.message : 'Unknown error'
-        onboardingLogger.warn(`Checkpoint validation failed: ${validationError}`, {
-          body: JSON.stringify(req.body)
-        })
-        return sendResponse<InvalidResponseMessage>(res, 400, {
-          error: validationError
-        })
-      }
-
-      try {
-        await onboarding.recordCheckpoint({
-          userIdentifier: checkpointReq.userIdentifier,
-          identifierType: checkpointReq.identifierType,
-          checkpointId: checkpointReq.checkpointId,
-          action: checkpointReq.action,
-          email: checkpointReq.email,
-          wallet: checkpointReq.wallet,
-          source: checkpointReq.source,
-          metadata: checkpointReq.metadata
-        })
-      } catch (e) {
-        onboardingLogger.error(
-          `[CP:${checkpointReq.checkpointId}][USER:${checkpointReq.userIdentifier}] Failed to record checkpoint: ${
-            isErrorWithMessage(e) ? e.message : 'Unknown error'
-          }`
-        )
-        return sendResponse<InvalidResponseMessage>(res, 500, { error: 'Internal server error' })
-      }
-
-      res.status(200).json({ success: true })
-    })
-
-    // Nudge queue dashboard — protected by ONBOARDING_API_KEY, safe for prod
-    app.get('/onboarding/pending-nudges', async (req: Request, res: Response) => {
-      const authHeader = req.headers['authorization'] ?? ''
-      const provided = authHeader.startsWith('Bearer ') ? authHeader.slice(7) : ''
-      if (provided !== onboardingApiKey) {
-        sendResponse(res, 401, { error: 'Unauthorized' })
-        return
-      }
-
-      try {
-        const sequences = [1, 2, 3] as const
-        const results: Record<string, { count: number; emails: string[] }> = {}
-
-        for (const seq of sequences) {
-          const nudges = await onboarding.getPendingNudges(seq)
-          // Group by checkpoint
-          const byCheckpoint = new Map<number, string[]>()
-          for (const n of nudges) {
-            const list = byCheckpoint.get(n.checkpointId) ?? []
-            list.push(n.email)
-            byCheckpoint.set(n.checkpointId, list)
-          }
-          for (const [cp, emails] of byCheckpoint) {
-            results[`CP${cp} - seq ${seq}`] = { count: emails.length, emails }
-          }
-        }
-
-        res.status(200).json(results)
-      } catch (e) {
-        const logger = logs.getLogger('onboarding-admin')
-        logger.error(`Failed to get pending nudges dashboard: ${isErrorWithMessage(e) ? e.message : 'Unknown error'}`)
-        res.status(500).json({ error: 'Failed to query pending nudges' })
-      }
-    })
-
-    // Admin endpoints — only mounted when ONBOARDING_ADMIN_ENABLED=true (local dev / staging)
-    const adminEnabled = (await config.getString('ONBOARDING_ADMIN_ENABLED')) === 'true'
-    if (adminEnabled) {
-      const adminLogger = logs.getLogger('onboarding-admin')
-      adminLogger.log('Admin endpoints enabled')
-
-      // Run the nudge evaluator manually (same as what cron does every 15 min)
-      app.post('/admin/onboarding/run-evaluator', async (_req: Request, res: Response) => {
-        try {
-          await nudgeJob.runEvaluator()
-          res.status(200).json({ success: true })
-        } catch (e) {
-          adminLogger.error(`Failed to run evaluator: ${isErrorWithMessage(e) ? e.message : 'Unknown error'}`)
-          res.status(500).json({ error: 'Failed to run evaluator' })
-        }
-      })
-
-      // Send a test nudge email directly (bypasses DB — just calls SendGrid)
-      app.post('/admin/onboarding/send-test-email', async (req: Request, res: Response) => {
-        const { to, checkpointId, sequence } = req.body
-
-        if (!to || !checkpointId || !sequence) {
-          sendResponse(res, 400, { error: 'Missing required fields: to, checkpointId, sequence' })
-          return
-        }
-
-        if (sequence < 1 || sequence > 3) {
-          sendResponse(res, 400, { error: 'sequence must be 1, 2, or 3' })
-          return
-        }
-
-        if (checkpointId < 1 || checkpointId > 7) {
-          sendResponse(res, 400, { error: 'checkpointId must be 1-7' })
-          return
-        }
-
-        try {
-          const messageId = await email.sendNudge({ to, checkpointId, sequence })
-          res.status(200).json({ success: true, messageId: messageId ?? null })
-        } catch (e) {
-          adminLogger.error(`Failed to send test email: ${isErrorWithMessage(e) ? e.message : 'Unknown error'}`)
-          res.status(500).json({ error: 'Failed to send email' })
-        }
-      })
-
-      // List pending nudges for a given sequence (what the cron would send next)
-      app.get('/admin/onboarding/pending-nudges/:sequence', async (req: Request, res: Response) => {
-        const sequence = parseInt(req.params.sequence, 10) as 1 | 2 | 3
-
-        if (![1, 2, 3].includes(sequence)) {
-          sendResponse(res, 400, { error: 'sequence must be 1, 2, or 3' })
-          return
-        }
-
-        try {
-          const nudges = await onboarding.getPendingNudges(sequence)
-          res.status(200).json({ sequence, count: nudges.length, nudges })
-        } catch (e) {
-          adminLogger.error(`Failed to get pending nudges: ${isErrorWithMessage(e) ? e.message : 'Unknown error'}`)
-          res.status(500).json({ error: 'Failed to query pending nudges' })
-        }
-      })
-    }
-
-    // PROM metrics endpoint
-    app.get(metricsPath, async (req: Request, res: Response) => {
-      if (metricsBearerToken) {
-        const authHeader = req.headers['authorization']
-        if (!authHeader) {
-          res.sendStatus(401)
-          return
-        }
-        const [, token] = authHeader.split(' ')
-        if (token !== metricsBearerToken) {
-          res.sendStatus(401)
-          return
-        }
-      }
-
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      const registry = (metrics as any).registry
-      if (!registry) {
-        res.sendStatus(500)
-        return
-      }
-
-      const body = await registry.metrics()
-      res.set('content-type', registry.contentType)
-      res.status(200).send(body)
-    })
-
-    Sentry.setupExpressErrorHandler(app)
-
-    server = new Server(httpServer, { cors: corsOptions })
-
-    server.on('connection', onConnection)
-
-    httpServer.listen(port)
-
-    logger.log(`Listening on port ${port}`)
-  }
-
-  const stop: IBaseComponent['stop'] = async () => {
-    if (!server) {
-      return
-    }
-    const logger = logs.getLogger('websocket-server')
-
-    logger.log('Stopping socket server...')
-
-    server.off('connection', onConnection)
-
-    await new Promise<void>(resolve => {
-      if (server) {
-        server.close(() => {
-          resolve()
-        })
-      }
-    })
-  }
-
-  return {
-    start,
-    stop
-  }
-}
